Incorporate USS

+++ scripts/zowe-explorer-ui-install.sh
@@ -43,4 +43,29 @@ pax -rf $EXPLORER_JES_PAX -ppx
 # Install explorer-jes ended
 #############################################
 
+
+#############################################
+# Install explorer-jes started
+EXPLORER_JES_PAX=$PWD/$(ls -t ./files/explorer-jes-*.pax | head -1)
+if [ ! -f $EXPLORER_JES_PAX ]; then
+  echo "JES Explorer UI (explorer-jes*.pax) missing"
+  echo "Installation terminated"
+  exit 0
+fi
+
+# NOTICE: zowe-install-iframe-plugin.sh will try to automatically create install folder based on plugin name
+EXPLORER_INSTALL_FOLDER=uss_explorer
+echo "Installing Explorer USS into ${ZOWE_ROOT_DIR}/${EXPLORER_INSTALL_FOLDER} ..."
+echo "Installing Explorer USS into ${ZOWE_ROOT_DIR}/${EXPLORER_INSTALL_FOLDER} ..."  >> $LOG_FILE
+umask 0002
+# remove old folders installed by atlas
+rm -fr "${ZOWE_ROOT_DIR}/${EXPLORER_INSTALL_FOLDER}" 2&>1 >/dev/null
+mkdir -p "${ZOWE_ROOT_DIR}/${EXPLORER_INSTALL_FOLDER}"
+# unpax package
+cd "${ZOWE_ROOT_DIR}/${EXPLORER_INSTALL_FOLDER}"
+echo "Unpax of ${EXPLORER_USS_PAX} into ${PWD}" >> $LOG_FILE
+pax -rf $EXPLORER_USS_PAX -ppx
+# Install explorer-uss ended
+#############################################
+
 echo "</zowe-explorer-ui-install.sh>" >> $LOG_FILE
